requirements.txt
keras==2.10.0
numpy==1.23.2
pandas==1.4.4
Pillow==9.2.0
streamlit_option_menu==0.3.2
